requirements.txt
aiogram==3.2.0
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
sqlalchemy==2.0.23